requirements.txt
absl-py==2.4.0
asttokens @ file:///C:/b/abs_9662ywy9fp/croot/asttokens_1743630464377/work
astunparse==1.6.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///C:/miniconda3/conda-bld/comm_1763119233835/work
contourpy==1.3.2
cycler==0.12.1
dotenv==0.9.9
debugpy @ file:///C:/miniconda3/conda-bld/debugpy_1762421657095/work
decorator @ file:///C:/miniconda3/conda-bld/decorator_1757341256088/work
easyocr==1.7.2
exceptiongroup @ file:///C:/miniconda3/conda-bld/exceptiongroup_1761560437674/work
executing @ file:///C:/miniconda3/conda-bld/executing_1757061252959/work
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.1
h5py==3.15.1
idna==3.11
ipykernel @ file:///C:/miniconda3/conda-bld/ipykernel_1762946093903/work
ipython @ file:///C:/b/abs_8eyhzleyrk/croot/ipython_1734548134403/work
itsdangerous==2.2.0
jedi @ file:///C:/b/abs_3a2kbnlclc/croot/jedi_1733987412687/work
Jinja2==3.1.6
jupyter_client @ file:///C:/Users/task_176839864521852/croot/jupyter_client_1768399093953/work
jupyter_core @ file:///C:/miniconda3/conda-bld/jupyter_core_1764588347534/work
keras==3.12.1
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline @ file:///C:/miniconda3/conda-bld/matplotlib-inline_1762779265381/work
mdurl==0.1.2
ml_dtypes==0.5.4
mysql-connector-python==9.6.0
namex==0.1.0
nest-asyncio @ file:///C:/b/abs_65d6lblmoi/croot/nest-asyncio_1708532721305/work
ninja==1.11.1.4
numpy==2.2.6
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.19.0
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049101700/work
parso @ file:///C:/miniconda3/conda-bld/parso_1762781925562/work
pillow==12.1.1
platformdirs @ file:///C:/miniconda3/conda-bld/platformdirs_1762356608986/work
prompt_toolkit @ file:///C:/miniconda3/conda-bld/prompt-toolkit_1761745021230/work
protobuf==6.33.5
psutil @ file:///C:/miniconda3/conda-bld/psutil_1761896539676/work
pure_eval @ file:///C:/miniconda3/conda-bld/pure_eval_1757067068391/work
pyclipper==1.3.0.post6
Pygments @ file:///C:/miniconda3/conda-bld/pygments_1762431425987/work
pyparsing==3.3.2
python-bidi==0.6.6
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pywin32 @ file:///C:/b/abs_6c9w_vp9zi/croot/pywin32_1754670534373/work
pyzmq @ file:///C:/miniconda3/conda-bld/pyzmq_1762375610449/work
requests==2.32.5
rich==14.3.3
shapely==2.0.7
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
stack_data @ file:///C:/miniconda3/conda-bld/stack_data_1757067051465/work
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
torch==2.6.0
torchvision==0.21.0
tornado @ file:///C:/Users/task_176599267555405/croot/tornado_1765992779662/work
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
typing_extensions==4.13.0
urllib3==2.6.3
wcwidth @ file:///C:/Users/task_176777900206875/croot/wcwidth_1767779217517/work
Werkzeug==3.1.6
wrapt==2.1.1
